requirements.txt
requests
xmltodict
python-dateutil
pytest==8.3.4
